Move the `get_name` function
